4.23: Blog List Expansion, step 11

+++ part4/controllers/blogs.js
@@ -1,6 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -8,16 +8,19 @@ blogsRouter.get('/', async (request, response) => {
   
   response.json(blogs)
   })
-  
-blogsRouter.post('/', async (request, response) => {
-  const { title, author, url, likes, userId } = request.body
+
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const { title, author, url, likes } = request.body
 
   if (!title || !url) {
     return response.status(400).json({ error: 'Title and URL are required' })
   }
 
-  const user = await User.findById(userId)
-
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({ error: 'User not authenticated' })
+  }
+  
   const blog = new Blog({
     title: title,
     author: author,
@@ -27,18 +30,28 @@ blogsRouter.post('/', async (request, response) => {
   })
 
   const savedBlog = await blog.save()
-
   user.blogs = user.blogs.concat(savedBlog._id);
   await user.save()
 
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
   const { id } = request.params
-  const result = await Blog.findByIdAndDelete(id)
+
+  const blog = await Blog.findById(id)
+
+  if (!request.user) {
+    return response.status(401).json({ error: 'Unauthorized' })
+  }
+
+  if ( blog.user.toString() !== request.user.id.toString()) {
+    return response.status(403).json({ error: 'Forbidden: You do not have permission to delete this blog' })
+  }
+
+  await Blog.findByIdAndDelete(id)
   response.status(204).end()
-})
+  })
 
 blogsRouter.put('/:id', async (request, response) => {
   const { id } = request.params
